微信支付接口

+++ app/src/main/java/com/gsq/iart/data/bean/PayConfigBean.kt
@@ -3,8 +3,8 @@ package com.gsq.iart.data.bean
 import java.io.Serializable
 
 data class PayConfigBean(
-    val id: Int,
-    val payAmount: Int,
+    val id: Long,
+    val payAmount: Float,
     val termDesc: String,
     val type: Int
 ) : Serializable
